Adding data validation for apis and fixing some bugs

+++ Backend/api/v1/User_Controller.php
@@ -1,5 +1,6 @@
 <?php
     require_once __DIR__ . "/../../models/User.php";
+    require_once __DIR__ . "/../../utilities/Controllers_Utilities.php";
     class User_Controller{
         // CRUD APIs Functions
         static function create(){
@@ -8,6 +9,8 @@
         }
         static function read(){
             $data = json_decode(file_get_contents("php://input"),true);
+            // if(!Controllers_Utilities::check_params($data,["id","username","email","password","title", "access_level","contact", "address"]))
+            //     return false;
             $user = User::read($data);
             echo json_encode([
                 "result" => boolval($user),
@@ -17,6 +20,8 @@
         }
         static function update(){
             $data = json_decode(file_get_contents("php://input"),true);
+            if(!Controllers_Utilities::check_params($data,["id","username","email","password","title", "access_level","contact", "address"]))
+                return false;
             $user = User::read_email($data);
             if($user && $user["id"] != $data['id']){
                 echo json_encode([
@@ -33,6 +38,8 @@
         }
         static function delete(){
             $data = json_decode(file_get_contents("php://input"),true);
+            if(!Controllers_Utilities::check_params($data,["id"]))
+                return false;
             $deleted = User::delete($data);
             echo json_encode([
                 "result" => $deleted,
@@ -42,6 +49,8 @@
         //
         static function login(){
             $data = json_decode(file_get_contents("php://input"),true);
+            if(!Controllers_Utilities::check_params($data,["email","password"]))
+                return false;
             $user = User::read($data);
             if ($user && password_verify($data['password'], $user['password'])) {
                 echo json_encode([
@@ -59,6 +68,8 @@
         }
         static function signup(){
             $data = json_decode(file_get_contents("php://input"),true);
+            if(!Controllers_Utilities::check_params($data,["username","email","password","title", "access_level","contact", "address"]))
+                return false;
             $user = User::read($data);
             if($user){
                 echo json_encode([
